+++ install_pydeps.py
@@ -31,7 +31,7 @@ from logging.config import fileConfig
 
 SBPIPE = os.environ["SBPIPE"]
 sys.path.append(os.path.join(SBPIPE, "sbpipe"))
-from sbpipe.sb_config import which
+from sbpipe.sbpipe_config import which
 
 
 def install_python_deps(requirements_file):
